app/requirements.txt
alembic==1.7.1
appdirs==1.4.4
asgiref==3.4.1
astroid==2.5.1
attrs==21.3.0
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
cryptography==35.0.0
cssselect==1.1.0
cssutils==2.3.0
distlib==0.3.2
Django==3.2.5
django-admin==1.0.2
email-to==0.1.0
fastapi==0.65.2
filelock==3.0.12
graphql-core==2.3.2
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
idna==3.3
image==1.5.33
importlib-resources==5.4.0
inflect==5.3.0
iniconfig==1.1.1
isort==5.7.0
joblib==1.0.1
lazy-object-proxy==1.5.2
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
Naked==0.1.31
neat-python==0.92
nltk==3.5
numpy==1.20.1
packaging==21.3
paramiko==2.8.0
passlib==1.7.4
Pillow==8.4.0
pluggy==1.0.0
premailer==3.10.0
promise==2.3
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.11.0
pydantic==1.8.2
pygame==2.0.1
PyJWT==2.1.0
pylint==2.7.2
PyNaCl==1.4.0
pyotp==2.6.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.18.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
qrcode==7.3
regex==2021.3.17
requests==2.26.0
Rx==1.6.1
shellescape==3.8.1
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.23
sqlalchemy-filters==0.12.0
sqlparse==0.4.1
starlette==0.14.2
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.14.0
virtualenv==20.4.7
watchgod==0.7
websockets==8.1
wrapt==1.12.1
zipp==3.6.0
